Add autentication

+++ app.py
@@ -1,14 +1,16 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request, jsonify, session
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import or_
 from werkzeug.security import generate_password_hash, check_password_hash
 from database import User, Message
+import os
 
 app = Flask(__name__)
+app.secret_key = os.urandom(24)
 engine = create_engine("sqlite:///messenger.db")
-session = sessionmaker(engine)
-db = session()
+db_session = sessionmaker(engine)
+db = db_session()
 
 
 @app.teardown_appcontext
@@ -17,8 +19,23 @@ def close_db(error):
     engine.dispose()
 
 
-@app.route('/user', methods=['POST'])
-def add_user():
+def get_current_user():
+    result = None
+
+    if "user" in session:
+        user = session["user"]
+        result = db.query(User).filter_by(user_id=user).first()
+
+    return result
+
+
+@app.route('/register', methods=['POST'])
+def register():
+    user = get_current_user()
+
+    if user:
+        return jsonify({"message": "The user is already logged in"})
+
     new_user_data = request.get_json()
 
     name = new_user_data['name']
@@ -36,147 +53,149 @@ def add_user():
 
     user_result = db.query(User).filter_by(email=email).first()
 
+    session["user"] = str(user_result.user_id)
+
     return jsonify({'user_id': user_result.user_id,
                     'name': user_result.name,
                     'email': user_result.email,
                     'password': user_result.password})
 
 
-@app.route('/user', methods=['GET'])
-def get_all_users():
-    all_users = db.query(User).all()
-
-    return_users = []
-
-    for user in all_users:
-        user_dict = {}
-        user_dict['user_id'] = user.user_id
-        user_dict['name'] = user.name
-        user_dict['email'] = user.email
-        user_dict['password'] = user.password
-
-        return_users.append(user_dict)
-
-    return jsonify({'users': return_users})
-
-
-@app.route('/user/<user_id>', methods=['GET'])
-def get_one_user(user_id):
-    user = db.query(User).filter_by(user_id=user_id).first()
+@app.route('/login', methods=['POST'])
+def login():
+    user = get_current_user()
 
     if user:
-        return jsonify({'user_id': user.user_id,
-                        'name': user.name,
-                        'email': user.email,
-                        'password': user.password})
+        return jsonify({"message": "The user is already logged in"})
 
-    return jsonify({'message': 'user not found'})
+    login_user_data = request.get_json()
 
+    email = login_user_data['email']
+    existing_user = db.query(User).filter_by(email=email).first()
 
-@app.route('/user/<user_id>', methods=['PUT'])
-def update_user(user_id):
-    user_to_update = db.query(User).filter_by(user_id=user_id).first()
-
-    new_user_data = request.get_json()
-
-    name = new_user_data['name']
-    email = new_user_data['email']
-    hashed_password = generate_password_hash(new_user_data['password'], method="sha256")
-
-    user_to_update.name = name
-    user_to_update.email = email
-    user_to_update.password = hashed_password
-    db.commit()
-
-    updated_user = db.query(User).filter_by(user_id=user_id).first()
-
-    return jsonify({'user_id': updated_user.user_id,
-                    'name': updated_user.name,
-                    'email': updated_user.email,
-                    'password': updated_user.password})
+    if existing_user:
+        if check_password_hash(existing_user.password, login_user_data['password']):
+            session["user"] = str(existing_user.user_id)
+            return jsonify({'user': {'user_id': existing_user.user_id,
+                                     'name': existing_user.name,
+                                     'email': existing_user.email,
+                                     'password': existing_user.password
+                                     }})
+        else:
+            return jsonify({'message': 'The password is wrong'})
+    else:
+        return jsonify({'message': 'The user not found'})
 
 
-@app.route('/user/<user_id>', methods=['DELETE'])
-def delete_user(user_id):
-    db.query(User).filter_by(user_id=user_id).delete()
-    db.commit()
-    return jsonify({'message': 'the user is deleted'})
+@app.route("/logout", methods=['POST'])
+def logout():
+    session.clear()
+    return jsonify({"message": "The user logged out"})
 
 
-@app.route('/user/<user_id>/message', methods=['POST'])
-def write_message(user_id):
+@app.route('/message', methods=['POST'])
+def write_message():
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
     new_message_data = request.get_json()
 
     receiver = new_message_data['receiver']
     message = new_message_data['message']
     subject = new_message_data['subject']
 
-    new_message = Message(sender=user_id, receiver=receiver,
+    if receiver == user.user_id:
+        return jsonify({'message': 'You cannot write letters to yourself'})
+
+    if len(message) == 0:
+        return jsonify(({'message': 'You have to write something in your message'}))
+
+    new_message = Message(sender=user.user_id, receiver=receiver,
                           message=message, subject=subject)
     db.add(new_message)
     db.commit()
 
-    message_result = db.query(Message).filter_by(sender=user_id).first()
+    message_result = db.query(Message).filter_by(sender=user.user_id).first()
 
-    return jsonify({'message_id': message_result.message_id,
-                    'sender': message_result.sender,
-                    'receiver': message_result.receiver,
-                    'message': message_result.message,
-                    'subject': message_result.subject,
-                    'create_date': message_result.create_date,
-                    'read': message_result.read
-                    })
+    return jsonify({'message': {'message_id': message_result.message_id,
+                                'sender': message_result.sender,
+                                'receiver': message_result.receiver,
+                                'message': message_result.message,
+                                'subject': message_result.subject,
+                                'create_date': message_result.create_date,
+                                'read': message_result.read
+                                }})
 
 
-@app.route('/user/<user_id>/message', methods=['GET'])
-def get_all_messages(user_id):
-    all_messages = db.query(Message).filter(or_(Message.sender==user_id, Message.receiver==user_id)).all()
+@app.route('/message', methods=['GET'])
+def get_all_messages():
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
+    all_messages = db.query(Message) \
+        .filter(or_(Message.sender == user.user_id, Message.receiver == user.user_id)).all()
 
     if len(all_messages) == 0:
-        return jsonify({'message': 'This user doesn\'t have messages yet.'})
+        return jsonify({'message': 'You don\'t have messages yet.'})
 
     return_messages = []
 
     for message in all_messages:
-        message_dict = {}
-        message_dict['message_id'] = message.message_id
-        message_dict['sender'] = message.sender
-        message_dict['receiver'] = message.receiver
-        message_dict['message'] = message.message
-        message_dict['subject'] = message.subject
-        message_dict['create_date'] = message.create_date
-        message_dict['read'] = message.read
+        message_dict = {'message_id': message.message_id,
+                        'sender': message.sender,
+                        'receiver': message.receiver,
+                        'message': message.message,
+                        'subject': message.subject,
+                        'create_date': message.create_date,
+                        'read': message.read
+                        }
 
         return_messages.append(message_dict)
 
     return jsonify({'messages': return_messages})
 
 
-@app.route('/user/<user_id>/message/unread', methods=['GET'])
-def get_unread_messages(user_id):
+@app.route('/message/unread', methods=['GET'])
+def get_unread_messages():
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
     unread_messages = db.query(Message) \
-        .filter(or_(receiver=user_id)) \
-        .filter(read=False).all()
+        .filter(Message.receiver == user.user_id) \
+        .filter(Message.read == False).all()
+
+    if len(unread_messages) == 0:
+        return jsonify({'message': 'You don\'t have unread messages'})
 
     return_messages = []
 
     for message in unread_messages:
-        message_dict = {}
-        message_dict['message_id'] = message.message_id
-        message_dict['sender'] = message.sender
-        message_dict['receiver'] = message.receiver
-        message_dict['message'] = message.message
-        message_dict['subject'] = message.subject
-        message_dict['create_date'] = message.create_date
-        message_dict['read'] = message.read
+        message_dict = {'message_id': message.message_id,
+                        'sender': message.sender,
+                        'receiver': message.receiver,
+                        'message': message.message,
+                        'subject': message.subject,
+                        'create_date': message.create_date,
+                        'read': message.read}
 
         return_messages.append(message_dict)
 
     return jsonify({'messages': return_messages})
 
 
-@app.route('/user/<user_id>/message/<message_id>', methods=['GET'])
-def get_one_message(user_id, message_id):
+@app.route('/message/<message_id>', methods=['GET'])
+def get_one_message(message_id):
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
     message = db.query(Message).filter_by(message_id=message_id).first()
 
     if message:
@@ -192,8 +211,13 @@ def get_one_message(user_id, message_id):
     return jsonify({'message': 'The message not found'})
 
 
-@app.route('/user/<user_id>/message/<message_id>', methods=['PUT'])
-def read_message(user_id, message_id):
+@app.route('/message/<message_id>', methods=['PUT'])
+def read_message(message_id):
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
     message_to_read = db.query(Message).filter_by(message_id=message_id).first()
 
     message_to_read.read = True
@@ -211,10 +235,16 @@ def read_message(user_id, message_id):
                                 }})
 
 
-@app.route('/user/<user_id>/message/<message_id>', methods=['DELETE'])
-def delete_message(user_id, message_id):
+@app.route('/message/<message_id>', methods=['DELETE'])
+def delete_message(message_id):
+    user = get_current_user()
+
+    if not user:
+        return jsonify({"message": "You have to login or register"})
+
     db.query(Message).filter_by(message_id=message_id).delete()
     db.commit()
+
     return jsonify({'message': 'the message is deleted'})
